Remove bootstrap file
